chore: unnuke metrics.go
